fix: move scripts used by compute_logs to their own directory to avoid shadowing dependencies
relates to https://github.com/dagster-io/dagster/issues/21920
